+++ CHANGES.rst
@@ -2,11 +2,13 @@ django-gm2m - changes
 =====================
 
 
-v0.1.1 (21-11-2014)
--------------------
+v0.2 (25-11-2014)
+-----------------
 
-Bugfix:
-- Inheritance of models with GM2MField no longer causes issues in Django 1.7
+Added:
+- Full Django 1.7+ migration support
+- through_fields option support
+- System checks (Django 1.7)
 
 
 v0.1 (08-10-2014)
@@ -18,6 +20,12 @@ Features:
 - Through models
 - Deletion behaviour customization (Django 1.6+)
 - Extended compatibility (Django 1.4 to 1.7, Python 2.6 to 3.4)
+
+v0.1.1 (21-11-2014)
+...................
+
+Bugfix:
+- Inheritance of models with GM2MField no longer causes issues in Django 1.7
 
 
 v0.0 (13-08-2014)